backend/requirements.txt
Django==6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
django-extensions==4.1
asgiref==3.11.0
sqlparse==0.5.5
PyJWT==2.10.1
python-decouple==3.8
